+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change log
 
-### 8.0.7 (09/23/2022 - )
+### 8.0.7 (09/23/2022 - 10/05/2022)
 
 * ext: updated dockerfile for security compliance
   > in order to enable kube "securityContext.runAsNonRoot: true", docker image should use numeric user (UID)